Improvements to the WebView
We set the viewport to ensure the view takes up the whole width, and we
disable scaling on double tap.  This seems to make the title go away in
the webviewpage, although I'm not sure.  Setting flickable: null may
help out, though it's not necessary and is commented right now.