Communication between frontend and WebView
QtWebKit.experimental has several functions for communicating between
QML and Javascript in the WebView.  I was hoping to use
postMessage/onMessageReceived, but I didn't get them to work.  So
instead, we have a message framework of evaluating javascript and
changing the document title.  This is mostly kept in separate libraries,
so we can swap them out in the future.

With this, we can pass the table of contents out to the QML level and
let selections there change the page of the book.