Allow opening a file from the command line
If this file isn't in the library, add it.

There's a silly work-around here, because qmlscene will grab any and all
arguments it can.  So we still all of the arguments to the shell script
in a single named argument that qmlscene doesn't know what to do with,
and so passes on to the running program.  Since we only want a single
filename now, it sorta works.  But if we want to do anything more
complicated, it's going to be a pain.