Change how margins are set
Now the user can only control the left and right margins, since all he
wants to do is set the text width to be comfortable.  But we ensure a
minimum of 1em around the content so there's a little space for it to
breathe.  Next, we'll set the top and bottom margins, and the default
side margins, based on the screen size.