Keep track of author display and sort separately
This allows us to take advantage of the opf:file-as attribute, when it
exists, to tell us how to sort authors, while still displaying the
author's name as indicated.

One problem: We should use attributeNS to get this attribute, in case
the namespace prefix isn't opf, but I couldn't get that to work.

Closes #3