Mitigate transparent header bug (#15)
If we don't set the top anchor of the perAuthorListView, it appears at
the right position vertically.  Therefore, we don't have to make it
flickable.  The main listview stays flickable and works correctly.  The
perAuthorListView is still visible through the header, but this list is
less likely to need scrolling and doesn't overlap with the header text.