Improvements to click packaging
The user's home directory should be specified as @{HOME} in the apparmor
profile, otherwise you'll get all manner of odd errors about parsing the
manifest file.  Which is of course a different file.  This is so obvious
that there's no need for it to be documented anywhere.

I don't think the icon is working yet, but that's okay since the icon is
ugly.