Remove Nemo Mobile FolderListModel
It's no longer in the Ubuntu Touch image.  (Yeah for not having a way to
note dependencies!)  We could use Qt.labs.folderlistmodel, but this
behaves somewhat differently and doesn't have a good way to tell when
it's ready for you to read the list of files.  So we've just rolled our
own solution in C++.

While we're removing stuff, we also got rid of the firstStart dialog,
which is no longer being used.