Implement cover view
Based on work by Stuart Langridge.  Ref #16.

We replace the recently-read list view with this, since that's a good
fit for the visual search.  Instead of using a single GridView for both
the covers and the lists, we swap between a GridView and a ListView.
Hopefully this will make things a little less tangled.  The positioning
of the GridView runs into the same sorts of problems as in #17, but the
work-around actually works in this case.

This includes a basic cover generator for books that don't have any.
Note that the title and author are added in QML, because doing flowed
text in SVG is difficult.  (Inkscape uses its own extension to do this,
but Qt doesn't recognize those elements.)