Remove WebViewLoader
It'd be nice to replace this with the new Ubuntu.Web 0.2, but that won't
be added to 14.04.  As a consolation, it'd be nice to replace this with
Ubuntu.Components.Extras.Browser 0.2, but that doesn't have download
support in 14.04 (though it is coming:
https://bugs.launchpad.net/oxide/+bug/1282063).  So we fall back to the
old QtWebKit-based browser, which we should be able to count upon now.
Refs #50.