Hold toolbar open by adjusting timeout
The MainView adjusts the layout when the toolbar is locked open, which
we don't want.  So instead, we turn off the timeout when the popover is
open, so the toolbar will stay open until the popover is closed.  In
anticipation of the new header, we do this only on the book page.
Closes #46.