Add texture to default book covers
We've made a tilable texture to use.  Ideally, we could use this as a
pattern, but the Qt SVG renderer doesn't support this.  So instead we do
enough tiling with clones to get the same effect.  The Qt renderer also
doesn't support linking external bitmaps, so we store the bitmap as
base64 encoded text, which we generate at compile time from a JPG file.

Closes #29.