Provide own Toolbar, rather than toggling deprecated toolbar
The ToolbarItems in BookPage are no longer assigned to "tools", so
MainView won't mess with them.  Instead, we manually set the "tools"
property of the Toolbar that we explicitly include.  Toolbar.qml comes
from the UITK, with an import line added and one change to account for
the fact that the tools won't be changed.

Closes #60.