Fix initial enablement of forward and back buttons on BookPage
These should be set by the initial history.clear() call, but that's not
working on 14.04.  Hooking into the onOpenedChanged handler worked
around this when we were using the built-in toolbar, but that's not
working with our custom toolbar.  Instead, just set both to be disabled
at the beginning, because we know that to be the correct state.