Smarter swipe-to-turn behavior
Previously, the direction of the page turn was determined by where the
swipe started, and whether it completed was determined by where the
swipe ended.  Now we pay attention to the direction that the swipe
starts.  Swipe to the left to go forward and to the right to go back.
Whether the turn completes depends on the release speed.  If it's above
a threshold, it completes if its in the same direction as it started and
aborts otherwise.  If the release speed is nearly zero, it completes if
there was net movement in the initial direction.

Taps still give instant page turns, with the direction determined by the
half on which the tap occurred.  Additionally, quick swipes lasting less
than 250 ms complete without any animation, as do longer swipes that end
with a large speed and a page turn.

Closes #34.