Improve behavior with long translated strings
We turn on wrapping for strings in the book settings dialog.  The layout
of the small buttons at the bottom will change if either of their
strings is detected to be too long.  The "Go" string was mistakenly
thought to be lenght limited; its button will actually expand if
necessary.

The "Auto" translation is problematic.  The translator comment has been
expanded to explain what it means in more detail.  We recomment an
em-dash if no translation will fit.