Add new bottom edge behavior for BookPage
This is based on the bottom edge from the Clock.  We add a set of
controls that come up first, with the full-page bottom edge fading in
beneath them.  The bottom edge element is not a page; we just fake that
look over top of the BookPage.  This way we avoid triggering the
visibility change handlers that are meant to notice when you've gone
back to the book list.