Preliminary support for CBZ files
There is no metadata provided at the moment, since I haven't found where
it's stored (or if it even is).  Right now, it is assumed that all the
files in the archive are images and the order in the archive is the
order in which they are to be presented.

Refs #78.