Only test ability to read in $HOME to determine confinement
With the download manager, we no longer have a need to write to the home
directory when running unconfined.  So, remove that capability from the
apparmor file and only test ability to read files in $HOME.

This is slightly trickier than you might think it is.  Running confined,
you can list the directory contents and read ~/.profile and ~/.bashrc.
But ~/.bash_logout is restricted, so we use that as our test case.