Support push imports from content hub
It should support multiple imports, but I haven't had a chance to test
that.  The import dialog probably won't be that useful in that case,
especially if there are errors.

The import dialog generaly doesn't show as soon as it should, since the
import processing happens before it has a chance to draw the dialog on
the screen.

Refs #61.